+++ src/cartridges/app_adyen_SFRA/cartridge/client/default/js/adyen_express/initializeCheckout.js
@@ -1,0 +1,16 @@
+async function initializeCheckout(paymentMethodsResponse) {
+  const applicationInfo = paymentMethodsResponse?.applicationInfo;
+  return window.AdyenWeb.AdyenCheckout({
+    environment: window.environment,
+    clientKey: window.clientKey,
+    locale: window.locale,
+    countryCode: window.countryCode,
+    analytics: {
+      analyticsData: { applicationInfo },
+    },
+  });
+}
+
+module.exports = {
+  initializeCheckout,
+};
