+++ src/cartridges/app_adyen_SFRA/cartridge/templates/default/product/components/addToCartButtonExtension.isml
@@ -1,24 +1,3 @@
-<isset name="AdyenHelper" value="${require('*/cartridge/adyen/utils/adyenHelper')}" scope="page"/>
-<isset name="AdyenConfigs" value="${require('*/cartridge/adyen/utils/adyenConfigs')}" scope="page" />
-<script src="${AdyenHelper.getCheckoutUrl()}" type="text/javascript"></script>
-<link rel="stylesheet" type="text/css" href="${AdyenHelper.getCheckoutCSS()}"/>
-<link rel="stylesheet" href="${URLUtils.staticURL('/css/adyenCss.css')}" />
-<script type="text/javascript">
-   window.environment = "${AdyenConfigs.getAdyenEnvironment()}";
-   window.clientKey = "${AdyenConfigs.getAdyenClientKey()}",
-   window.locale = "${request.locale}";
-   window.getPaymentMethodsURL = "${URLUtils.https('Adyen-GetPaymentMethods')}";
-   window.shippingMethodsUrl = "${URLUtils.https('Adyen-ShippingMethods')}";
-   window.selectShippingMethodUrl = "${URLUtils.https('Adyen-SelectShippingMethod')}";
-   window.paymentsDetailsURL = "${URLUtils.https('Adyen-PaymentsDetails')}";
-   window.paymentFromComponentURL = "${URLUtils.https('Adyen-PaymentFromComponent')}";
-   window.showConfirmationAction = "${URLUtils.https('Adyen-ShowConfirmationPaymentFromComponent')}";
-   window.returnUrl = "${URLUtils.https('Cart-Show')}";
-   window.createTemporaryBasketUrl = "${URLUtils.https('Adyen-CreateTemporaryBasket')}";
-   window.isApplePayExpressOnPdpEnabled = "${AdyenConfigs.isApplePayExpressOnPdpEnabled()}";
-   window.isGooglePayExpressOnPdpEnabled = "${AdyenConfigs.isGooglePayExpressOnPdpEnabled()}";
-   window.fractionDigits = "${AdyenHelper.getFractionDigits()}"
-</script>
-
+<isinclude template="adyen/adyenExpressMetadata" />
 <input type="hidden" id="adyen-token" name="${pdict.csrf.tokenName}" value="${pdict.csrf.token}"/>
 <div id="express-payment-buttons"></div>
